fixed msvc and watcom build in the same directory
msvc for some reason would prefer to link imago.lib etc from the project
directory, instead of the Debug/Release directories which are specified
in the linker search paths. This would conflict with building with
watcom in the same directory, which leaves libraries in the project
root. Changed the linker command line to explicitly link
$(OutDir)\foo.lib to avoid this issue.